Linked the Backend and FrontEnd for the timeline Profile
Created the connection between the BackEnd and the FrontEnd for the population of the timelineProfile.

I added some comments in different files explaining what is happening, if you don't know yet how it's working check them out.

I added also other comments on the things that we can improove to make the code easier to understand.